+++ lib/rubocop/config_loader.rb
@@ -174,6 +174,11 @@ module RuboCop
 
       def resolve_inheritance_from_gems(hash, gems)
         (gems || {}).each_pair do |gem_name, config_path|
+          if gem_name == 'rubocop'
+            fail ArgumentError,
+                 "can't inherit configuration from the rubocop gem"
+          end
+
           hash['inherit_from'] = Array(hash['inherit_from'])
           # Put gem configuration first so local configuration overrides it.
           hash['inherit_from'].unshift gem_config_path(gem_name, config_path)